+++ decoder/LAVAudio/parser/parser.h
@@ -20,7 +20,7 @@
 
 #pragma once
 
-static inline const uint8_t *find_marker32_position(const uint8_t *pBuffer, unsigned uBufSize, uint32_t marker)
+static inline const uint8_t *find_marker32_position(const uint8_t *pBuffer, size_t uBufSize, uint32_t marker)
 {
   uint32_t state = 0;
   for (unsigned i = 0; i < uBufSize; ++i) {
